Move massless tests
