+++ newton/_src/solvers/kamino/tests/test_solver_kamino.py
@@ -41,6 +41,10 @@ from newton.tests.utils import basics
 
 wp.set_module_options({"enable_backward": False, "default_grid_stride": False})
 
+BASE_HEIGHT = 0.5
+SIM_DT = 1.0 / 60.0
+UNIT_INERTIA = wp.mat33(0.1, 0.0, 0.0, 0.0, 0.1, 0.0, 0.0, 0.0, 0.1)
+
 ###
 # Kernels
 ###
@@ -1458,6 +1462,98 @@ class TestSolverKaminoImpl(unittest.TestCase):
         np.testing.assert_allclose(multi_state_n.u_i.numpy(), multi_final_u_i, rtol=rtol, atol=atol)
         np.testing.assert_allclose(multi_state_n.q_j.numpy(), multi_final_q_j, rtol=rtol, atol=atol)
         np.testing.assert_allclose(multi_state_n.dq_j.numpy(), multi_final_dq_j, rtol=rtol, atol=atol)
+
+
+class TestSolverKaminoMasslessBodies(unittest.TestCase):
+    """Massless bodies are supported when welded to the world and rejected otherwise."""
+
+    def test_massless_body_welded_to_world_stays_at_rest(self):
+        """Verify a massless body welded to the world stays at rest, along with its welded child.
+
+        This is the ``world -> fixed -> base -> fixed -> link0`` topology produced by importing a
+        fixed-base URDF whose root link carries no inertial properties.
+        """
+        builder = newton.ModelBuilder(up_axis=newton.Axis.Z)
+        base_xform = wp.transform(wp.vec3(0.0, 0.0, BASE_HEIGHT), wp.quat_identity())
+        base = builder.add_link(xform=base_xform, mass=0.0, label="base")
+        link0 = builder.add_link(xform=base_xform, mass=2.0, inertia=UNIT_INERTIA, label="link0")
+        joint_base = builder.add_joint_fixed(parent=-1, child=base, parent_xform=base_xform)
+        joint_link0 = builder.add_joint_fixed(parent=base, child=link0)
+        builder.add_articulation([joint_base, joint_link0])
+        model = builder.finalize()
+
+        inv_mass = model.body_inv_mass.numpy()
+        self.assertEqual(inv_mass[base], 0.0)
+        self.assertGreater(inv_mass[link0], 0.0)
+
+        solver = SolverKamino(model)
+        state_0, state_1 = model.state(), model.state()
+        control = model.control()
+        newton.eval_fk(model, model.joint_q, model.joint_qd, state_0)
+
+        # A spurious velocity on the massless base is unrecoverable, so it must be caught per step
+        # rather than only at the end, where it would have grown far beyond the tolerance.
+        for step in range(30):
+            state_0.clear_forces()
+            solver.step(state_0, state_1, control, contacts=None, dt=SIM_DT)
+            state_0, state_1 = state_1, state_0
+            np.testing.assert_allclose(
+                state_0.body_qd.numpy()[[base, link0]],
+                0.0,
+                atol=1.0e-5,
+                err_msg=f"welded chain gained velocity at step {step}",
+            )
+
+        np.testing.assert_allclose(state_0.body_q.numpy()[[base, link0]][:, 2], BASE_HEIGHT, atol=1.0e-5)
+
+    def test_massless_tip_behind_revolute_joint_is_rejected(self):
+        """Verify a massless tip that is not welded to the world is rejected at construction.
+
+        Kamino cannot simulate this model.
+
+        The tip reaches the world only through a revolute joint, so it is free to move yet cannot be
+        accelerated by every constraint reaction. Its missing response would propagate through the
+        weld and prevent physically meaningful motion of the massive link carrying it, so the model
+        must be rejected rather than simulated.
+        """
+        builder = newton.ModelBuilder(up_axis=newton.Axis.Z)
+        link_xform = wp.transform(wp.vec3(0.0, 0.0, BASE_HEIGHT), wp.quat_identity())
+        link0 = builder.add_link(xform=link_xform, mass=2.0, inertia=UNIT_INERTIA, label="link0")
+        tip = builder.add_link(xform=link_xform, mass=0.0, label="tip")
+        joint_link0 = builder.add_joint_revolute(parent=-1, child=link0, axis=newton.Axis.Y)
+        joint_tip = builder.add_joint_fixed(parent=link0, child=tip)
+        builder.add_articulation([joint_link0, joint_tip])
+        model = builder.finalize()
+
+        with self.assertRaises(ValueError) as ctx:
+            SolverKamino(model)
+
+        message = str(ctx.exception)
+        self.assertIn("singular inertial properties", message)
+        self.assertIn("'tip'", message)
+        self.assertNotIn("'link0'", message)
+
+    def test_partially_singular_tip_behind_revolute_joint_is_rejected(self):
+        """Verify a partially singular tip is rejected at construction."""
+        builder = newton.ModelBuilder(up_axis=newton.Axis.Z)
+        link_xform = wp.transform(wp.vec3(0.0, 0.0, BASE_HEIGHT), wp.quat_identity())
+        link0 = builder.add_link(xform=link_xform, mass=2.0, inertia=UNIT_INERTIA, label="link0")
+        tip = builder.add_link(xform=link_xform, mass=2.0, inertia=UNIT_INERTIA, label="tip")
+        joint_link0 = builder.add_joint_revolute(parent=-1, child=link0, axis=newton.Axis.Y)
+        joint_tip = builder.add_joint_fixed(parent=link0, child=tip)
+        builder.add_articulation([joint_link0, joint_tip])
+        model = builder.finalize()
+
+        inv_inertia = model.body_inv_inertia.numpy()
+        inv_inertia[tip] = np.diag(np.array([0.0, 1.0, 1.0], dtype=np.float32))
+        model.body_inv_inertia.assign(inv_inertia)
+
+        with self.assertRaises(ValueError) as ctx:
+            SolverKamino(model)
+
+        message = str(ctx.exception)
+        self.assertIn("singular inverse inertia", message)
+        self.assertIn("'tip'", message)
 
 
 ###
